*: replace Grpc.Core with Grpc.Net.Client library
The generated code works with both libraries. `Grpc.Core` is the
original gRPC library implemented in C. This is already deprecated in
favor of `Grpc.Net.Client`, a new implementation written purely in C#.

The commit also changes the target framework from netcoreapp3.1 to
netstandard2.1 which is less specific and should increase the project
compatibility.

The `Grpc.Core` dependency is still required to run the Auth tests to
setup the testing gRPC server.