build: use NETStandard.Library and add System.Text.Json
The System.Text.Json library is available un the .NET Core framework but
after switching to .NET Standard we need to import it as a dependency.

The previous release failed because we were still using .NET Core in the
Bazel build, so the PR checks did not fail. This commit also fixes that.