build: bump to .NET 8.0 and build tools to latest
We were using a pretty old version of .NET, and updating it required
bumping the rules_dotnet build tool to the latest version. This has
forced us to apply lots of changes to the build and the release scripts.

The main benefit is that the latest rules_dotnet version includes better
support to download NuGet dependencies. After this change we no longer
need to manually update the nuget_deps.bzl script.